+++ hololinked/server/__init__.py
@@ -1,9 +1,11 @@
 __version__ = "0.1"
 
 # Order of import is reflected in this file to avoid circular imports
+from .constants import HTTP_METHODS
 from .config import *
 from .serializers import *
 from .zmq_message_brokers import *
+from .events import *
 from .decorators import *
 from .remote_parameter import *
 from .database import *
