+++ hololinked/server/zmq_message_brokers.py
@@ -14,8 +14,7 @@ from enum import Enum
 from .utils import create_default_logger, run_method_somehow, wrap_text
 from .config import global_config
 from .constants import JSON, ZMQ_PROTOCOLS, ServerTypes
-from .serializers import (JSONSerializer, PickleSerializer, BaseSerializer, 
-                        SerpentSerializer, serializers, MsgpackSerializer)
+from .serializers import (JSONSerializer, BaseSerializer, serializers, MsgpackSerializer)
                         # DillSerializer, 
 from ..param.parameterized import Parameterized
 
@@ -1791,7 +1790,7 @@ class MessageMappedZMQClientPool(BaseZMQClient):
 
     async def async_execute(self, instance_name : str, instruction : str, arguments : typing.Dict[str, typing.Any] = EMPTY_DICT, 
                     *, context : typing.Dict[str, typing.Any] = EMPTY_DICT, raise_client_side_exception = False, 
-                    invokation_timeout : typing.Optional[float] = 3, execution_timeout : typing.Optional[float] = None,
+                    invokation_timeout : typing.Optional[float] = 5, execution_timeout : typing.Optional[float] = None,
                     argument_schema : typing.Optional[JSON] = None) -> typing.Dict[str, typing.Any]:
         """
         sends message and receives reply.
@@ -1809,7 +1808,7 @@ class MessageMappedZMQClientPool(BaseZMQClient):
             see execution context definitions
         raise_client_side_exceptions: bool, default False
             raise exceptions from server on client side
-        invokation_timeout: float, default 3 
+        invokation_timeout: float, default 5
             server side timeout
         execution_timeout: float, default None
             client side timeout, not the same as timeout passed to server, recommended to be None in general cases. 
@@ -1834,57 +1833,49 @@ class MessageMappedZMQClientPool(BaseZMQClient):
         """
         self.stop_poll = True
 
-    # async def ping_all_servers(self):
-    #     """
-    #     ping all servers connected to the client pool
-    #     """
-    #     replies = await asyncio.gather(*[self.async_execute(
-    #         instance_name, '/ping') for instance_name in self.pool.keys()]) # type: typing.List[typing.Dict[str, typing.Any]]
-    #     sorted_reply = dict() 
-    #     for reply, instance_name in zip(replies, self.pool.keys()):
-    #         sorted_reply[instance_name] = reply.get("returnValue", False if reply.get("exception", None) is None else True)  # type: ignore
-    #     return sorted_reply
+    async def async_execute_in_all(self, instruction : str, instance_names : typing.Optional[typing.List[str]] = None,  
+                        arguments : typing.Dict[str, typing.Any] = EMPTY_DICT, context : typing.Dict[str, typing.Any] = EMPTY_DICT, 
+                        invokation_timeout : typing.Optional[float] = 5, execution_timeout : typing.Optional[float] = None, 
+                        raise_client_side_exception = False) -> typing.Dict[str, typing.Any]:
+        """
+        execute a specified instruction in all instance_names
+        """
+        if not instance_names:
+            instance_names = self.pool.keys()
+        gathered_replies = await asyncio.gather(*[
+            self.async_execute(instance_name, instruction, arguments, 
+                        context=context, raise_client_side_exception=raise_client_side_exception,
+                        invokation_timeout=invokation_timeout, execution_timeout=execution_timeout) 
+                for instance_name in instance_names])
+        replies = dict()
+        for instance_name, reply in zip(instance_names, gathered_replies):
+            replies[instance_name] = reply
+        return replies  
+    
+    async def async_execute_in_all_remote_objects(self, instruction : str, arguments : typing.Dict[str, typing.Any] = EMPTY_DICT, 
+                        context : typing.Dict[str, typing.Any] = EMPTY_DICT, 
+                        invokation_timeout : typing.Optional[float] = 5, execution_timeout : typing.Optional[float] = None, 
+                        raise_client_side_exception = False) -> typing.Dict[str, typing.Any]:
+        return await self.async_execute_in_all(instruction=instruction, 
+                        instance_names=[instance_name for instance_name, client in self.pool.items() if client.server_type == ServerTypes.REMOTE_OBJECT],
+                        arguments=arguments, context=context, invokation_timeout=invokation_timeout, 
+                        execution_timeout=execution_timeout, raise_client_side_exception=raise_client_side_exception)
+    
+    async def async_execute_in_all_eventloops(self, instruction : str, arguments : typing.Dict[str, typing.Any] = EMPTY_DICT, 
+                        context : typing.Dict[str, typing.Any] = EMPTY_DICT, 
+                        invokation_timeout : typing.Optional[float] = 5, execution_timeout : typing.Optional[float] = None, 
+                        raise_client_side_exception = False) -> typing.Dict[str, typing.Any]:
+        return await self.async_execute_in_all(instruction=instruction, 
+                        instance_names=[instance_name for instance_name, client in self.pool.items() if client.server_type == ServerTypes.EVENTLOOP],
+                        arguments=arguments, context=context, invokation_timeout=invokation_timeout, 
+                        execution_timeout=execution_timeout, raise_client_side_exception=raise_client_side_exception)
 
-    # def organised_gathered_replies(self, instance_names : List[str], gathered_replies : List[Any], context : Dict[str, Any] = EMPTY_DICT):
-    #     """
-    #     First thing tomorrow
-    #     """
-    #     plain_reply = context.pop('plain_reply', False)
-    #     if not plain_reply:
-    #         replies = dict(
-    #             returnValue = dict(),
-    #             state = dict()
-    #         )
-    #         for  instance_name, reply in zip(instance_names, gathered_replies):
-    #             replies["state"].update(reply["state"])
-    #             replies["returnValue"][instance_name] = reply.get("returnValue", reply.get("exception", None))
-    #     else: 
-    #         replies = {}
-    #         for instance_name, reply in zip(instance_names, gathered_replies):
-    #             replies[instance_name] = reply
-    #     return replies  
-
-    # async def async_execute_in_all(self, instruction : str, arguments : Dict[str, Any] = EMPTY_DICT, 
-    #                     context : Dict[str, Any] = EMPTY_DICT, raise_client_side_exception = False) -> Dict[str, Any]:
-    #     instance_names = self.pool.keys()
-    #     gathered_replies = await asyncio.gather(*[
-    #         self.async_execute(instance_name, instruction, arguments, context, raise_client_side_exception) for instance_name in instance_names])
-    #     return self.organised_gathered_replies(instance_names, gathered_replies, context)
-    
-    # async def async_execute_in_all_remote_objects(self, instruction : str, arguments : Dict[str, Any] = EMPTY_DICT,
-    #                             context : Dict[str, Any] = EMPTY_DICT, raise_client_side_exception = False) -> Dict[str, Any]:
-    #     instance_names = [instance_name for instance_name, client in self.pool.items() if client.server_type == ServerTypes.USER_REMOTE_OBJECT]
-    #     gathered_replies = await asyncio.gather(*[
-    #         self.async_execute(instance_name, instruction, arguments, context, raise_client_side_exception) for instance_name in instance_names])
-    #     return self.organised_gathered_replies(instance_names, gathered_replies, context)
-    
-    # async def async_execute_in_all_eventloops(self, instruction : str, arguments : Dict[str, Any] = EMPTY_DICT, 
-    #                     context : Dict[str, Any] = EMPTY_DICT, raise_client_side_exception = False) -> Dict[str, Any]:
-    #     instance_names = [instance_name for instance_name, client in self.pool.items() if client.server_type == ServerTypes.EVENTLOOP]
-    #     gathered_replies = await asyncio.gather(*[
-    #         self.async_execute(instance_name, instruction, arguments, context, raise_client_side_exception) for instance_name in instance_names])
-    #     return self.organised_gathered_replies(instance_names, gathered_replies, context)
-    
+    async def ping_all_servers(self):
+        """
+        ping all servers connected to the client pool
+        """
+        return await self.async_execute_in_all(instruction='/ping')
+        
     def __contains__(self, name : str) -> bool:
         return name in self.pool
 
@@ -1942,150 +1933,5 @@ class AsyncioEventPool:
 
 
 
-class Event:
-
-    def __init__(self, name : str, URL_path : typing.Optional[str] = None) -> None:
-        self.name = name 
-        # self.name_bytes = bytes(name, encoding = 'utf-8')
-        self.URL_path = URL_path or '/' + name
-        self._unique_identifier = None # type: typing.Optional[str]
-        self._owner = None  # type: typing.Optional[Parameterized]
-        # above two attributes are not really optional, they are set later. 
-
-    @property
-    def owner(self):
-        return self._owner        
-        
-    @property
-    def publisher(self) -> "EventPublisher": 
-        return self._publisher
-    
-    @publisher.setter
-    def publisher(self, value : "EventPublisher") -> None:
-        if not hasattr(self, '_publisher'):
-            self._publisher = value
-            self._publisher.register_event(self)
-        else:
-            raise AttributeError("cannot reassign publisher attribute of event {}".format(self.name)) 
-
-    def push(self, data : typing.Any = None, serialize : bool = True):
-        self.publisher.publish_event(self._unique_identifier, data, serialize)
-
-
-
-class CriticalEvent(Event):
-
-    def __init__(self, name : str, message_broker : BaseZMQServer, acknowledgement_timeout : float = 3, 
-                URL_path : typing.Optional[str] = None) -> None:
-        super().__init__(name, URL_path)
-        self.message_broker = message_broker
-        self.aknowledgement_timeout = acknowledgement_timeout
-
-    def push(self, data : typing.Any = None):
-        super().push(data)
-        self.message_broker.receive_acknowledgement()
-
-
-
-class EventPublisher(BaseZMQServer):
-
-    def __init__(self,  identity : str, context : typing.Union[zmq.Context, None] = None, **serializer) -> None:
-        super().__init__(server_type=ServerTypes.UNKNOWN_TYPE, **serializer)
-        self.context = context or zmq.Context()
-        self.identity = identity
-        self.socket = self.context.socket(zmq.PUB)
-        for i in range(1000, 65535):
-            try:
-                self.socket_address = "tcp://127.0.0.1:{}".format(i)
-                self.socket.bind(self.socket_address)
-            except zmq.error.ZMQError as ex:
-                if ex.strerror.startswith('Address in use'):
-                    pass 
-                else:
-                    print("Following error while atttempting to bind to socket address : {}".format(self.socket_address))
-                    raise ex from None
-            else:
-                self.logger = self.get_logger(identity, "PUB", "TCP", logging.DEBUG)
-                self.logger.info("created event publishing socket at {}".format(self.socket_address))
-                break
-        self.events = set() # type: typing.Set[Event] 
-        self.event_ids = set() # type: typing.Set[bytes]
-
-    def register_event(self, event : Event) -> None:
-        # unique_str_bytes = bytes(unique_str, encoding = 'utf-8') 
-        if event._unique_identifier in self.events:
-            raise AttributeError(
-                "event {} already found in list of events, please use another name.".format(event.name),
-                "Also, Remotesubobject and RemoteObject cannot share event names."
-            )
-        self.event_ids.add(event._unique_identifier)
-        self.events.add(event) 
-        self.logger.info("registered event '{}' serving at PUB socket with address : {}".format(event.name, self.socket_address))
-               
-    def publish_event(self, unique_str : bytes, data : typing.Any, serialize : bool = True) -> None: 
-        if unique_str in self.event_ids:
-            if serialize:
-                self.socket.send_multipart([unique_str, self.json_serializer.dumps(data)])
-            else:
-                self.socket.send_multipart([unique_str, data])
-        else:
-            raise AttributeError("event name {} not yet registered with socket {}".format(unique_str, self.socket_address))
-        
-    def exit(self):
-        if not hasattr(self, 'logger'):
-            self.logger = create_default_logger('{}|{}'.format(self.__class__.__name__, uuid4()))
-        try:
-            self.socket.close(0)
-            self.logger.info("terminated event publishing socket with address '{}'".format(self.socket_address))
-        except Exception as E:
-            self.logger.warn("could not properly terminate context or attempted to terminate an already terminated context at address '{}'. Exception message : {}".format(
-                self.socket_address, str(E)))
-        try:
-            self.context.term()
-            self.logger.info("terminated context of event publishing socket with address '{}'".format(self.socket_address)) 
-        except Exception as E: 
-            self.logger.warn("could not properly terminate socket or attempted to terminate an already terminated socket of event publishing socket at address '{}'. Exception message : {}".format(
-                self.socket_address, str(E)))
-      
-
-class EventConsumer(BaseZMQClient):
-
-    def __init__(self, event_URL : str, socket_address : str, identity : str, client_type = HTTP_SERVER, **kwargs) -> None:
-        super().__init__(server_address=b'unknown', server_instance_name='unkown', client_type=client_type, **kwargs)
-        # Prepare our context and publisher
-        self.event_URL = bytes(event_URL, encoding = 'utf-8')
-        self.socket_address = socket_address
-        self.identity = identity
-        self.context = zmq.asyncio.Context()
-        self.socket = self.context.socket(zmq.SUB)
-        self.socket.connect(socket_address)
-        self.socket.setsockopt(zmq.SUBSCRIBE, self.event_URL)
-        self.logger = self.get_logger(identity, self.socket_address, logging.DEBUG, self.__class__.__name__)
-        self.logger.info("connected event consuming socket to address {}".format(self.socket_address))
-      
-    async def receive_event(self, deserialize = False):
-        _, contents = await self.socket.recv_multipart()
-        if not deserialize: 
-            return contents
-        return self.json_serializer.loads(contents)
-    
-    def exit(self):
-        if not hasattr(self, 'logger'):
-            self.logger = create_default_logger('{}|{}'.format(self.__class__.__name__, uuid4()))
-        try:
-            self.socket.close(0)
-            self.logger.info("terminated event consuming socket with address '{}'".format(self.socket_address))
-        except Exception as E:
-            self.logger.warn("could not properly terminate context or attempted to terminate an already terminated context at address '{}'. Exception message : {}".format(
-                self.socket_address, str(E)))
-        try:
-            self.context.term()
-            self.logger.info("terminated context of event consuming socket with address '{}'".format(self.socket_address)) 
-        except Exception as E: 
-            self.logger.warn("could not properly terminate socket or attempted to terminate an already terminated socket of event consuming socket at address '{}'. Exception message : {}".format(
-                self.socket_address, str(E)))
-            
-
-
 __all__ = ['AsyncZMQServer', 'AsyncPollingZMQServer', 'ZMQServerPool', 'RPCServer',
-           'SyncZMQClient', 'AsyncZMQClient', 'MessageMappedZMQClientPool', 'Event', 'CriticalEvent']
+           'SyncZMQClient', 'AsyncZMQClient', 'MessageMappedZMQClientPool']
